+++ src/main/java/io/tarantool/driver/core/RequestFutureManager.java
@@ -1,36 +1,11 @@
 package io.tarantool.driver.core;
 
-import io.tarantool.driver.api.TarantoolClientConfig;
 import io.tarantool.driver.mappers.MessagePackValueMapper;
 import io.tarantool.driver.protocol.TarantoolRequest;
 
-import java.util.Map;
 import java.util.concurrent.CompletableFuture;
-import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.ScheduledExecutorService;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.TimeoutException;
 
-/**
- * Keeps track of submitted requests, finishing them by timeout and allowing asynchronous request processing
- *
- * @author Alexey Kuzin
- */
-public class RequestFutureManager implements AutoCloseable {
-    private final ScheduledExecutorService timeoutScheduler;
-    private final TarantoolClientConfig config;
-    private final Map<Long, TarantoolRequestMetadata> requestFutures = new ConcurrentHashMap<>();
-
-    /**
-     * Basic constructor.
-     *
-     * @param config           tarantool client configuration
-     * @param timeoutScheduler scheduled executor for handling request timeouts
-     */
-    public RequestFutureManager(TarantoolClientConfig config, ScheduledExecutorService timeoutScheduler) {
-        this.config = config;
-        this.timeoutScheduler = timeoutScheduler;
-    }
+public interface RequestFutureManager extends AutoCloseable {
 
     /**
      * Submit a request ID for tracking. Provides a {@link CompletableFuture} for tracking the request completion.
@@ -41,9 +16,7 @@ public class RequestFutureManager implements AutoCloseable {
      * @param <T>          target response body type
      * @return {@link CompletableFuture} that completes when a response is received from Tarantool server
      */
-    public <T> CompletableFuture<T> submitRequest(TarantoolRequest request, MessagePackValueMapper resultMapper) {
-        return submitRequest(request, config.getRequestTimeout(), resultMapper);
-    }
+    <T> CompletableFuture<T> submitRequest(TarantoolRequest request, MessagePackValueMapper resultMapper);
 
     /**
      * Submit a request ID for tracking. Provides a {@link CompletableFuture} for tracking the request completion.
@@ -55,21 +28,9 @@ public class RequestFutureManager implements AutoCloseable {
      * @param <T>            target response body type
      * @return {@link CompletableFuture} that completes when a response is received from Tarantool server
      */
-    public <T> CompletableFuture<T> submitRequest(TarantoolRequest request,
-                                                  int requestTimeout,
-                                                  MessagePackValueMapper resultMapper) {
-        CompletableFuture<T> requestFuture = new CompletableFuture<>();
-        long requestId = request.getHeader().getSync();
-        requestFuture.whenComplete((r, e) -> requestFutures.remove(requestId));
-        requestFutures.put(requestId, new TarantoolRequestMetadata(requestFuture, resultMapper));
-        timeoutScheduler.schedule(() -> {
-            if (!requestFuture.isDone()) {
-                requestFuture.completeExceptionally(new TimeoutException(String.format(
-                        "Failed to get response for request id: %d within %d ms", requestId, requestTimeout)));
-            }
-        }, requestTimeout, TimeUnit.MILLISECONDS);
-        return requestFuture;
-    }
+    <T> CompletableFuture<T> submitRequest(TarantoolRequest request,
+                                           int requestTimeout,
+                                           MessagePackValueMapper resultMapper);
 
     /**
      * Get a request me instance bound to the passed request ID
@@ -77,12 +38,7 @@ public class RequestFutureManager implements AutoCloseable {
      * @param requestId ID of a request to Tarantool server (sync ID)
      * @return {@link CompletableFuture} that completes when a response is received from Tarantool server
      */
-    public TarantoolRequestMetadata getRequest(Long requestId) {
-        return requestFutures.get(requestId);
-    }
+    TarantoolRequestMetadata getRequest(Long requestId);
 
-    @Override
-    public void close() {
-        requestFutures.values().forEach(f -> f.getFuture().join());
-    }
+    void close();
 }
